# Optimise queries search for a chain of OR strings