@ironage 2019-03-04
Here's the actual benchmark results for reference (100000 rows, 1000 query conditions):

Before:
```
QueryChainedOrStrings (MemOnly, EncryptionOff):   min 353.80ms     max 497.77ms     median 399.02ms     avg 394.69ms     stddev  44.38ms
QueryChainedOrStrings (MemOnly, EncryptionOn):    min 351.54ms     max 456.18ms     median 368.16ms     avg 375.97ms     stddev  32.92ms
QueryChainedOrStrings (Full   , EncryptionOff):   min 350.13ms     max 360.58ms     median 351.87ms     avg 352.84ms     stddev   3.24ms
QueryChainedOrStrings (Full   , EncryptionOn):    min 350.49ms     max 400.45ms     median 352.64ms     avg 357.29ms     stddev  15.26ms

test/benchmark-common-tasks/realm-benchmark-common-tasks  28.96s user 0.22s system 94% cpu 30.743 total
```
After:
```
QueryChainedOrStrings (MemOnly, EncryptionOff):     min      5ms (-98.59%)           max   7.27ms (-98.54%)           med   5.77ms (-98.55%)           avg   5.81ms (-98.53%)           stddev   658us (-98.52%)
QueryChainedOrStrings (MemOnly, EncryptionOn):      min   4.92ms (-98.60%)           max   5.77ms (-98.74%)           med   5.19ms (-98.59%)           avg   5.20ms (-98.62%)           stddev   242us (-99.26%)
QueryChainedOrStrings (Full   , EncryptionOff):     min   4.87ms (-98.61%)           max   5.38ms (-98.51%)           med   4.95ms (-98.59%)           avg      5ms (-98.58%)           stddev   134us (-95.88%)
QueryChainedOrStrings (Full   , EncryptionOn):      min   4.93ms (-98.59%)           max   5.43ms (-98.64%)           med   5.21ms (-98.52%)           avg   5.20ms (-98.54%)           stddev   158us (-98.97%)

test/benchmark-common-tasks/realm-benchmark-common-tasks  0.81s user 0.16s system 42% cpu 2.271 total
```