@ironage 2019-03-05
It's a good point, there is an overhead cost of computing a string hash. The majority of time is spent computing our custom StringData hash (not sure how performant it tries to be). That being the case, I added a simple loop check to iterate through the conditions and check for matches without hashing anything. Based on the following tests, I found that the threshold for choosing this over the hash is around 20 conditions.


  | no optimisations | unordered_set (hashing) | list search
-- | -- | -- | --
1 |   |   |  
2 | 0.7 | 1.8 | 0.3
3 | 1 | 3.5 | 0.4
4 | 1.4 | 4 | 0.7
5 | 1.7 | 4 | 0.8
10 | 5.4 | 4 | 1.9
20 | 10.5 | 4 | 3.5
30 | 16 | 3.6 | 5
50 | 25.5 | 3.2 | 8.5


<img width="532" alt="screenshot 2019-03-04 at 17 15 18" src="https://user-images.githubusercontent.com/2826060/53773902-1e44e880-3ea1-11e9-945a-910ebb1f6982.png">